Fixed norm errors
